# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IAEDU API",
        "description": "API para el sistema educativo IAEDU",
        "contact": {
            "email": "admin@iaedu.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/alerts": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Obtener lista de alertas",
                "operationId": "5caf81e921e803b0f06d5242d3e62d56",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Filtrar por estudiante",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tipo",
                        "in": "query",
                        "description": "Filtrar por tipo de alerta",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtrar por estado (activa, resuelta)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "Incluir relaciones (student)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de alertas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Alerts retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Alert"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Crear nueva alerta",
                "operationId": "a827893cea07afe23ecf2d6d39873929",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_id",
                                    "tipo",
                                    "mensaje"
                                ],
                                "properties": {
                                    "student_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "academica"
                                    },
                                    "mensaje": {
                                        "type": "string",
                                        "example": "Bajo rendimiento académico"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "activa",
                                            "resuelta"
                                        ],
                                        "example": "activa"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alerta creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Alert created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Alert"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/alerts/{id}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Obtener alerta específica",
                "operationId": "7108f67ad863bde7f166ec91650f7ad0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la alerta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "Incluir relaciones (student)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alerta obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Alert retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Alert"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alerta no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Actualizar alerta",
                "operationId": "f5e4118632ed032cfaaccb635029df9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la alerta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "student_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "academica"
                                    },
                                    "mensaje": {
                                        "type": "string",
                                        "example": "Bajo rendimiento académico"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "activa",
                                            "resuelta"
                                        ],
                                        "example": "activa"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alerta actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Alert updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Alert"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alerta no encontrada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Eliminar alerta",
                "operationId": "893ce20da0a3664628f0f6e28ccef3d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la alerta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alerta eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Alert deleted successfully"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alerta no encontrada"
                    }
                }
            }
        },
        "/api/alerts/statistics": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Obtener estadísticas de alertas",
                "operationId": "f6e388dbabcc13f6757b6a5d5a1bd7ea",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Filtrar por estudiante",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estadísticas obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Alert statistics retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_alerts": {
                                                    "type": "integer"
                                                },
                                                "active_alerts": {
                                                    "type": "integer"
                                                },
                                                "resolved_alerts": {
                                                    "type": "integer"
                                                },
                                                "alerts_by_type": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Alert": {
                "title": "Alert",
                "description": "Modelo de alerta del sistema",
                "properties": {
                    "id": {
                        "description": "ID único de la alerta",
                        "type": "integer",
                        "example": 1
                    },
                    "student_id": {
                        "description": "ID del estudiante",
                        "type": "integer",
                        "example": 1
                    },
                    "tipo": {
                        "description": "Tipo de alerta",
                        "type": "string",
                        "example": "academica"
                    },
                    "mensaje": {
                        "description": "Mensaje de la alerta",
                        "type": "string",
                        "example": "Bajo rendimiento académico"
                    },
                    "estado": {
                        "description": "Estado de la alerta",
                        "type": "string",
                        "enum": [
                            "activa",
                            "resuelta"
                        ],
                        "example": "activa"
                    },
                    "fecha": {
                        "description": "Fecha de la alerta",
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}